Rename gen-markdown-index-2
