+++ app/shell/py/pie/setup.py
@@ -23,7 +23,6 @@ setup(
             'render-jinja-template=pie.render_jinja_template:main',
             'render-study-json=pie.render_study_json:main',
             'include-filter=pie.include_filter:main',
-            'gen-markdown-index=pie.gen_markdown_index:main',
             'process-yaml=pie.process_yaml:main',
             'detect-html-dicts=pie.detect_html_dicts:main',
         ],
