+++ src/gen-markdown-index/index.yml
@@ -1,6 +1,6 @@
 title: gen-markdown-index
 name: gen-markdown-index
 id: dist_gen_markdown_index
-gen-markdown-index-2:
+gen-markdown-index:
   link: false
   show: false
